Fixed some MSVC compiler warnings

+++ Emulator/Base/CmdQueueTypes.h
@@ -191,7 +191,7 @@ struct CmdTypeEnum : util::Reflection<CmdType, CmdType> {
 typedef struct
 {
     Option option;
-    isize value;
+    i64 value;
     isize id;
 }
 ConfigCmd;
